% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{evo.model}
\alias{evo.model}
\title{Set the evolutionary model parameters and pass out a list of 3 lists}
\usage{
evo.model(type, sub, rate)
}
\value{
List of 3 lists with prior distributions parameters
}
\description{
Set the evolutionary model parameters and pass out a list of 3 lists
}

