% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{load.nex}
\alias{load.nex}
\title{Import nexus file}
\usage{
load.nex(align)
}
\arguments{
\item{align}{String of Nexus file.}
}
\value{
Loaded nexus file as matrix of base pairs/amino acids/codons for each taxa.
}
\description{
Import nexus file
}
\examples{
data <- load.nex("primates.nex")
}

