% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{phylo.run}
\alias{phylo.run}
\title{Run Bayesian Inference}
\usage{
phylo.run(aligned, evo.par, N, K)
}
\arguments{
\item{aligned}{Imported Nexus file matrix}

\item{evo.par}{Output of the evo.model}

\item{N}{number of taxa}

\item{K}{the length of the sequences}
}
\value{
Shiny stan output of the posterior distributions for the different parameters
}
\description{
Run Bayesian Inference
}

