\name{phylostan-package}
\alias{phylostan-package}
\alias{phylostan}
\docType{package}
\title{
Using stan to inference phylogenies on small datasets.
}
\description{
Coms4761 project attempt to implement a newer algorithm to to use Bayesian inference to construct phylogentic trees. The package relies on aligned .nex files and outputs statistics and diagnostics for determining the best tree. 
}
\details{
\tabular{ll}{
Package: \tab phylostan\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-05-11\cr
}
Phylostan accepts data from Nexus files and then allows the user to specify the evolutionary model parameters:
Type: DNA, Codon, Amino Acid
Substitution Model: GTR, Juke-Cantor

}
\author{
Ian Walshe
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
