% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{evo.model}
\alias{evo.model}
\title{Set the evolutionary model parameters and pass out a list of 3 lists}
\usage{
evo.model(type, sub)
}
\arguments{
\item{type}{String codon, amino, dna}

\item{sub}{String gtr, jc are the two implmentations right now}
}
\value{
List of parameters to be incorporated into model priors
}
\description{
Set the evolutionary model parameters and pass out a list of 3 lists
}
\examples{
data <- evo.model("dna","jc")
}

