\name{phylostan-package}
\alias{phylostan-package}
\alias{phylostan}
\docType{package}
\title{
Using stan to inference phylogenies on smalld datasets. (short line)
}
\description{
Coms4761 project attempt to implement a newer algorithm to to use Bayesian inference to construct phylogentic trees. The package relies on aligned .nex files and outputs statistics and diagnostics for determining the best tree. 
}
\details{
\tabular{ll}{
Package: \tab phylostan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-28\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Ian Walshe
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
